Big refactoring with DX12Device
This should not change the behavior of the engine. This is just refactoring.

Made the DX12Device globally accessible, instead of passing it by arguments. The device was already global to begin with (in main.cpp)
This makes code simpler and nicer, without having to carry it from function to function. This also has the nice advantage of being able to use it in descructors/Release functions (in next CL)