We can render things again!
Added a new descriptor heap per command list that is shader visible. This means that we need to copy descriptors from a non shader visible heap to a shader visible heap.
Right now this is just a very big hack, since we still need a nice way to bind resources.

+ Added FullScreen renderer with DrawUtils (still a nasty hack for now)
* Removed GPU handle ownership from TextureDX12. The reason is that GPU handles should only be used with shader visible heaps. This heap in our case is temporary, so that would mean the CPU/GPU handle of the resource are in different heaps, not good.