Robustness part 3
* Replaced pointers to references when applicable. The new rule is:
Can it be null? Yes: pointer. No: reference
Exception: Function returning an allocated object, even if it can't be null: pointer

Why? The goal is to make the user of a function/class aware that hey, this returns a pointer so, maybe it can be null, or maybe this allocates an object on the heap.
Or this function takes a reference, so I'm not allowed to pass null.

* Missed a few files in previous commit